+++ app/Http/Controllers/Panel/PlaceController.php
@@ -61,6 +61,12 @@ class PlaceController extends Controller
 
   public function destroy($id)
   {
+
+    if (!($request->cookie(env("PRIVATE_ACCESS_COOKIE_NAME")) == env("PRIVATE_ACCESS_COOKIE_ADMIN_VALUE"))) {
+		  $url = '/panel/prohibitedaction/'.str_replace("/", "%20", $request->url());
+		  return redirect($url);
+    }
+
 		$place = Place::findOrFail($id);
 		$place->delete();
