+++ app/Http/Controllers/Panel/RecipeController.php
@@ -101,6 +101,12 @@ class RecipeController extends Controller
 
   public function destroy($id)
   {
+
+    if (!($request->cookie(env("PRIVATE_ACCESS_COOKIE_NAME")) == env("PRIVATE_ACCESS_COOKIE_ADMIN_VALUE"))) {
+		  $url = '/panel/prohibitedaction/'.str_replace("/", "%20", $request->url());
+		  return redirect($url);
+    }
+
     $recipe = Recipe::findOrFail($id);
     $recipe->delete();
