+++ app/Http/Controllers/Panel/SuggestionController.php
@@ -55,6 +55,12 @@ class SuggestionController extends Controller
 
   public function destroy($id)
   {
+
+    if (!($request->cookie(env("PRIVATE_ACCESS_COOKIE_NAME")) == env("PRIVATE_ACCESS_COOKIE_ADMIN_VALUE"))) {
+		  $url = '/panel/prohibitedaction/'.str_replace("/", "%20", $request->url());
+		  return redirect($url);
+    }
+
 		$suggestion = Suggestion::findOrFail($id);
 		$suggestion->delete();
