+++ app/Http/Controllers/Panel/UnitController.php
@@ -57,6 +57,12 @@ class UnitController extends Controller
 
   public function destroy($id)
   {
+
+    if (!($request->cookie(env("PRIVATE_ACCESS_COOKIE_NAME")) == env("PRIVATE_ACCESS_COOKIE_ADMIN_VALUE"))) {
+		  $url = '/panel/prohibitedaction/'.str_replace("/", "%20", $request->url());
+		  return redirect($url);
+    }
+
 		$unit = Unit::findOrFail($id);
 		$unit->delete();
